+++ TidalPy/utilities/classes/base.py
@@ -1,4 +1,4 @@
-from ..._version import version
+from ... import version
 
 
 class TidalPyClass: